Были удалены лишние файлы, Program перенесен в MyReceipts.App( так же изменен MyReceipts.App.csproj)
