+++ MyReceipts.App/MyReceipts.App.csproj
@@ -15,4 +15,8 @@
       <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.6" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\MyReceipts.UI\MyReceipts.UI.csproj" />
+    </ItemGroup>
+
 </Project>
