+++ MyReceipts.App/Program.cs
@@ -1,3 +1,16 @@
-﻿// See https://aka.ms/new-console-template for more information
-
-Console.WriteLine("Hello, World!");
+﻿var app = new AppUI();
+app.StartMenu();
+var userChoice = app.ChoiceMenu();
+while (userChoice != "2. Выход")
+{
+    /*switch (userChoice)
+    {
+        case "1. Поиск рецепта по ингредиентам":
+            app.SearchByIngridients();
+            var favSystemChoice = app.FavouriteSystem();
+            app.FavouriteWindow(favSystemChoice, userChoice);
+            break;
+        case "2. Выход":
+            break;
+    }*/
+}
